# java/socialapp/src/main/java/com/contoso/socialapp/dto/LikeRequest.java
package com.contoso.socialapp.dto;

import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.AllArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class LikeRequest {
    private String username;
}