# java/socialapp/src/main/java/com/contoso/socialapp/dto/UpdateCommentRequest.java
package com.contoso.socialapp.dto;

import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.AllArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UpdateCommentRequest {
    private String username;
    private String content;
}